+++ src/app/(dashboard)/scanner/page.tsx
@@ -1,0 +1,6 @@
+import React from 'react'
+import ScannerClient from './ScannerClient'
+
+export default function ScannerPage() {
+  return <ScannerClient />
+}